% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseb_expansion.R
\name{baseb_expansion}
\alias{baseb_expansion}
\title{Adds the next base-b element to an existing base-b sequence}
\usage{
baseb_expansion(ain, b)
}
\arguments{
\item{ain}{Either a an array containing an existing base-b expansion, or a scalar integer indicating the length for a new base-b expansion}

\item{b}{Base for integer expansions used in the sequence}
}
\value{
An expanded base-b expansion
}
\description{
Adds the next base-b element to an existing base-b sequence
}
