% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mesh_size.R
\name{calc_mesh_size}
\alias{calc_mesh_size}
\title{Calculates the number of points in a mesh of fineness phi, covering a hypercube in d dimensions}
\usage{
calc_mesh_size(phi, d)
}
\arguments{
\item{phi}{The scalar fineness of the mesh}

\item{d}{The number of dimensions for the unit spheroid}
}
\value{
A list of: corner_pts - Count of points extreme (+/- 1) in all dim, edge_pts - Count of points extreme in all but one dimen, face_pts - Count of points extreme in all but two dimen, total_pts - Sum of:  corner_pts + edge_pts + face_pts
}
\description{
Calculates the number of points in a mesh of fineness phi, covering a hypercube in d dimensions
}
