% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_at_origin.R
\name{center_at_origin}
\alias{center_at_origin}
\title{Creates a new ellipsoid object equivalent to the given hyperellipsoid (hellipse), but centered at the origin.}
\usage{
center_at_origin(hellip)
}
\arguments{
\item{hellip}{The original object, to be shifted}
}
\value{
list of two: hellip2 - the re-centered hyperellipsoid and mu - the amount of the translation
}
\description{
Creates a new ellipsoid object equivalent to the given hyperellipsoid (hellipse), but centered at the origin.
}
