% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Show basic facts of a hyperellipsoid (center, shape, size)}
\usage{
display(hellip)
}
\arguments{
\item{hellip}{A valid hyperellipsoid object}
}
\value{
none
}
\description{
Show basic facts of a hyperellipsoid (center, shape, size)
}
\examples{
display(hellip)
}
