% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_ellipsoid.R
\name{distance_to_ellipsoid}
\alias{distance_to_ellipsoid}
\title{Calculates the distance between the given point and the given hyperellipsoid (hellipse), measured along a line connecting the point with the center of the ellipsoid.  If the given point is interior to the ellipsoid, then the reported distance is negative.   If the point is on the ellipsoid, the distance is zero. If the point lies outside the ellipsoid, the distance is postive.}
\usage{
distance_to_ellipsoid(hellip, point)
}
\arguments{
\item{hellip}{The shape to which distance is measured}

\item{point}{Where distance is measured from}
}
\value{
The Euclidean distance between the point and the corresponding spot on the surface of hellip
}
\description{
Calculates the distance between the given point and the given hyperellipsoid (hellipse), measured along a line connecting the point with the center of the ellipsoid.  If the given point is interior to the ellipsoid, then the reported distance is negative.   If the point is on the ellipsoid, the distance is zero. If the point lies outside the ellipsoid, the distance is postive.
}
