% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldcount.R
\name{fieldcount}
\alias{fieldcount}
\title{Utility function for number of fields in a hyperellipsoid}
\usage{
fieldcount(hellip)
}
\arguments{
\item{hellip}{A valid hyperellipsoid object}
}
\value{
The number of fields in the object (should be 3)
}
\description{
Utility function for number of fields in a hyperellipsoid
}
\examples{
fieldcount(hellip)
}
