% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_adj_2Dface.R
\name{fill_adj_2Dface}
\alias{fill_adj_2Dface}
\title{Creates a phi x phi grid (i.e., the mesh on a single two-dimensional face of a larger hypercube) of d-dimensional points, where the regularity of the grid has been adjusted to avoid  clustering in the corners.}
\usage{
fill_adj_2Dface(d, phi)
}
\arguments{
\item{d}{The number of dimensions for the unit spheroid}

\item{phi}{Fineness of the mesh along each dimension of the 2D face}
}
\value{
A phi x phi x d array of points.  The points (each facemesh2D(i,j,:)) are identically equal to one in the first d-2 dimensions, so that the mesh varies only in the final two dimensions.
}
\description{
Creates a phi x phi grid (i.e., the mesh on a single two-dimensional face of a larger hypercube) of d-dimensional points, where the regularity of the grid has been adjusted to avoid  clustering in the corners.
}
