% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_adj_2Dface_beta.R
\name{fill_adj_2Dface_beta}
\alias{fill_adj_2Dface_beta}
\title{Calculates the factor, beta in [0, 1], that interpolates the pth equidistant point between the two endpoints, z_one and z_phi, for and adjusted 2D mesh of fineness phi in d dimensions.}
\usage{
fill_adj_2Dface_beta(p, phi, z_one, z_phi)
}
\arguments{
\item{p}{...}

\item{phi}{Fineness of the mesh along each dimension of the 2D face}

\item{z_one}{...}

\item{z_phi}{...}
}
\value{
beta
}
\description{
Calculates the factor, beta in [0, 1], that interpolates the pth equidistant point between the two endpoints, z_one and z_phi, for and adjusted 2D mesh of fineness phi in d dimensions.
}
