% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Get hyperellipsoid property from the specified object and return the value.
Property names are:
center, shape, and size}
\usage{
get(hellip, propName)
}
\arguments{
\item{hellip}{A valid hyperellipsoid object}

\item{propName}{A string of the desired property}
}
\value{
The value of the indicated property
}
\description{
Get hyperellipsoid property from the specified object and return the value.
Property names are:
center, shape, and size
}
