% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperellipsoid.R
\name{hyperellipsoid}
\alias{hyperellipsoid}
\title{Hyperellipsoid class constructor}
\usage{
hyperellipsoid(...)
}
\arguments{
\item{...}{mu - The vector for the center point,
sig - The matrix determining the shape; for elliptical probability distributions, sig will be the inverse dispersion matrix,
c - The scalar determining the size}
}
\value{
A new hyperellipsoid object
}
\description{
Hyperellipsoid class constructor
}
\examples{
hyperellipsoid()
}
