% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corners.R
\name{make_corners}
\alias{make_corners}
\title{Fills a mesh (corn_mesh) with d-dimensional points representing all corners of a d-dimensional cube encompassing a d-dimensional unit spheroid.}
\usage{
make_corners(d, normalize)
}
\arguments{
\item{d}{The number of dimensions for the unit spheroid}

\item{normalize}{Whether to scale the corner points onto the sphere or not}
}
\value{
A d x 2^d array of corner points
}
\description{
Fills a mesh (corn_mesh) with d-dimensional points representing all corners of a d-dimensional cube encompassing a d-dimensional unit spheroid.
}
