% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{make_edges}
\alias{make_edges}
\title{Fills a mesh with d-dimensional points representing all non-corner edge points of a d-dimensional cube encompassing a  d-dimensional unit spheroid.}
\usage{
make_edges(d, phi, normalize)
}
\arguments{
\item{d}{The number of dimensions for the unit spheroid}

\item{phi}{Fineness of the mesh along the edge (i.e., the total number of points, "including" the corners)}

\item{normalize}{Whether to scale the corner points onto the sphere or not}
}
\value{
A d x d*2^(d-1)*(phi-2) array of edge points
}
\description{
Fills a mesh with d-dimensional points representing all non-corner edge points of a d-dimensional cube encompassing a  d-dimensional unit spheroid.
}
