% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_faces.R
\name{make_faces}
\alias{make_faces}
\title{Fills a mesh with d-dimensional points representing all non-edge face points of a d-dimensional cube encompassing a  d-dimensional unit spheroid.}
\usage{
make_faces(d, phi, normalize)
}
\arguments{
\item{d}{The number of dimensions for the unit spheroid}

\item{phi}{Fineness of the mesh along each dimension of the 2D face}

\item{normalize}{Whether to scale the corner points onto the sphere or not}
}
\value{
A d x d*(d-1)*2^(d-3)*(phi-2)^2 array of face points
}
\description{
Fills a mesh with d-dimensional points representing all non-edge face points of a d-dimensional cube encompassing a  d-dimensional unit spheroid.
}
