% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_baseb_expansion.R
\name{new_baseb_expansion}
\alias{new_baseb_expansion}
\title{Creates a new base-b sequence of a designated length}
\usage{
new_baseb_expansion(k, b)
}
\arguments{
\item{k}{The integer to expand}

\item{b}{Base for integer expansions used in the sequence}
}
\value{
The expansion of the integer k
}
\description{
Creates a new base-b sequence of a designated length
}
