% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_to_coordaxes.R
\name{rotate_to_coordaxes}
\alias{rotate_to_coordaxes}
\title{Rotates the ellipsoid (hellip) so its principal axes align with the coordinate axes.  Both ellipsoids are centered at the origin.  Note that there are (2^d)*d! valid ways to rotate the ellipsoid to the axes. This algorithm does not prescribe which solution will be provided.}
\usage{
rotate_to_coordaxes(hellip)
}
\arguments{
\item{hellip}{The shape to be rotated, must be centered at the origin}
}
\value{
A list of: hellip2 - A new hyperellipsoid, rotated to the coordinate axes and tfm - the transformation matrix that creates the rotation
}
\description{
Rotates the ellipsoid (hellip) so its principal axes align with the coordinate axes.  Both ellipsoids are centered at the origin.  Note that there are (2^d)*d! valid ways to rotate the ellipsoid to the axes. This algorithm does not prescribe which solution will be provided.
}
