% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeparam_normal_distn.R
\name{sizeparam_normal_distn}
\alias{sizeparam_normal_distn}
\title{Calculates the size paramater for a d-dimensional hyperellipsoid conforming to a normal (i.e., Gaussian) distribution.}
\usage{
sizeparam_normal_distn(prob, d)
}
\arguments{
\item{prob}{The target probability threshold}

\item{d}{Number of dimensions in the multivariate distribution}
}
\value{
The appropriate (scalar) size parameter
}
\description{
Calculates the size paramater for a d-dimensional hyperellipsoid conforming to a normal (i.e., Gaussian) distribution.
}
\examples{
sizeparam_normal_distn(0.95, 6)
}
