% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeparam_t_distn.R
\name{sizeparam_t_distn}
\alias{sizeparam_t_distn}
\title{Calculates the size paramater for a d-dimensional hyperellipsoid conforming to a Student's t distribution.}
\usage{
sizeparam_t_distn(prob, d, nu)
}
\arguments{
\item{prob}{The target probability threshold}

\item{d}{Number of dimensions in the multivariate distribution}

\item{nu}{Degrees of freedom parameter for the t distribution}
}
\value{
The appropriate (scalar) size parameter
}
\description{
Calculates the size paramater for a d-dimensional hyperellipsoid conforming to a Student's t distribution.
}
\examples{
sizeparam_t_distn(0.95, 6, 5)
}
