% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid_mesh.R
\name{spheroid_mesh}
\alias{spheroid_mesh}
\title{Generates a Cartesian mesh of d-dimensional scenarios based on the given ellipsoid.  This function does not assume that the ellipsoid is centered at the origin.}
\usage{
spheroid_mesh(d, phi, normalize)
}
\arguments{
\item{d}{The number of dimensions for the unit spheroid}

\item{phi}{The scalar fineness of the mesh}

\item{normalize}{Whether to normalize points from the cube onto the sphere or not (TRUE/FALSE)}
}
\value{
A d x N array with each column a scenario
}
\description{
Generates a Cartesian mesh of d-dimensional scenarios based on the given ellipsoid.  This function does not assume that the ellipsoid is centered at the origin.
}
