% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch_to_unitspheroid.R
\name{stretch_to_unitspheroid}
\alias{stretch_to_unitspheroid}
\title{Stretches the ellipsoid (hellip) to the unit spheroid of the same dimension.  Both the input ellipsoid and unit spheroid are centered at the origin.}
\usage{
stretch_to_unitspheroid(hellip)
}
\arguments{
\item{hellip}{The original shape to be stretched}
}
\value{
A list of: hellip1 - a new unit spheroid, mapped from the ellipsoid and tfm - transformation matrix that creates the stretching
}
\description{
Stretches the ellipsoid (hellip) to the unit spheroid of the same dimension.  Both the input ellipsoid and unit spheroid are centered at the origin.
}
