% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_ellipsoid.R
\name{transform_ellipsoid}
\alias{transform_ellipsoid}
\title{Applies the given linear transformation, tfm, to the given ellipsoid.  The ellipsoid (hellip) must be centered at the origin.}
\usage{
transform_ellipsoid(hellip, tfm)
}
\arguments{
\item{hellip}{The original shape to be transformed}

\item{tfm}{A d x d linear transformation matrix}
}
\value{
A transformed ellipsoid, centered at the origin
}
\description{
Applies the given linear transformation, tfm, to the given ellipsoid.  The ellipsoid (hellip) must be centered at the origin.
}
