% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vandercorput.R
\name{vandercorput}
\alias{vandercorput}
\title{Generates a base-b Van der Corput sequence with k elements}
\usage{
vandercorput(k, b)
}
\arguments{
\item{k}{Number of elements in the sequence}

\item{b}{Base for integer expansions used in the sequence}
}
\value{
k-dimensional array containing the sequence
}
\description{
Generates a base-b Van der Corput sequence with k elements
}
